+++ src/main/java/payroll/OrderController.java
@@ -21,6 +21,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+
 @RestController
 class OrderController {
 
@@ -130,7 +131,9 @@ class OrderController {
   
   @DeleteMapping("/orders/{id}")
   ResponseEntity<?> delete(@PathVariable Long id) {
-	    orderRepository.deleteById(id);
-	    return ResponseEntity.noContent().build();
+	  if(orderRepository.existsById(id)) 
+	       orderRepository.deleteById(id);
+	  else throw new OrderNotFoundException(id) ;
+	  return new ResponseEntity<>("Order id: "+id+" deleted", HttpStatus.OK);
   }
 }
